+++ modules/@apostrophecms/util/index.js
@@ -5,7 +5,7 @@
 
 // ## Options
 //
-// ### `logger`
+// ### `logger` (LEGACY)
 //
 // A function which accepts `apos` and returns an object with
 // at least `info`, `debug`, `warn` and `error` methods. These methods should
@@ -16,6 +16,11 @@
 // `apos.util.error`, etc. are routed through this object
 // by Apostrophe. This provides compatibility out of
 // the box with many popular logging modules, including `winston`.
+// NOTE: this option is deprecated, you should configure `@apostrophecms/log`
+// module instead. This option will still work for BC reasons, but switching
+// to structured logging and module `self.logInfo()`, `self.logError()`, etc
+// is highly recommended.
+// Read more in the`@apostrophecms/log` module documentation.
 
 const _ = require('lodash');
 const he = require('he');
@@ -45,6 +50,17 @@ module.exports = {
     self.regExpQuote = require('regexp-quote');
     self.warnedDev = {};
     return self.enableLogger();
+  },
+  handlers(self) {
+    return {
+      'apostrophe:destroy': {
+        async destroyLogger() {
+          if (self.logger.destroy) {
+            await self.logger.destroy();
+          }
+        }
+      }
+    };
   },
   methods(self) {
     return {
@@ -504,7 +520,20 @@ module.exports = {
         }
       },
       enableLogger() {
-        self.logger = self.options.logger ? self.options.logger(self.apos) : require('./lib/logger.js')(self.apos);
+        // Legacy, configured via this module.
+        if (self.options.logger) {
+          self.logger = self.options.logger(self.apos);
+          return;
+        }
+        // New, configured via the `log` module.
+        const logOpts = self.apos.structuredLog.options;
+        if (logOpts.logger) {
+          self.logger = typeof logOpts.logger === 'function'
+            ? logOpts.logger(self.apos)
+            : logOpts.logger;
+          return;
+        }
+        self.logger = require('./lib/logger.js')(self.apos);
       },
       // Log a message. The default
       // implementation wraps `console.log` and passes on
@@ -517,11 +546,12 @@ module.exports = {
       // If the logger has no `log` method, the `info` method
       // is used. This allows an instance of `bole` or similar
       // to be used directly.
-      log(msg) {
+      log(...args) {
+        // kept for bc
         if (!self.logger.log) {
-          return self.logger.info.apply(self.logger.info, arguments);
+          return self.logger.info(...self.convertLegacyLogPayload(args));
         }
-        self.logger.log.apply(self.logger, arguments);
+        self.logger.log(...self.convertLegacyLogPayload(args));
       },
       // Log an informational message. The default
       // implementation wraps `console.info` and passes on
@@ -530,8 +560,8 @@ module.exports = {
       // Overrides should be written with support for
       // substitution strings in mind. See the
       // `console.log` documentation.
-      info(msg) {
-        self.logger.info.apply(self.logger, arguments);
+      info(...args) {
+        self.logger.info(...self.convertLegacyLogPayload(args));
       },
       // Log a debug message. The default implementation wraps
       // `console.debug` if available, otherwise `console.log`,
@@ -540,8 +570,8 @@ module.exports = {
       // Overrides should be written with support for
       // substitution strings in mind. See the
       // `console.warn` documentation.
-      debug(msg) {
-        self.logger.debug.apply(self.logger, arguments);
+      debug(...args) {
+        self.logger.debug(...self.convertLegacyLogPayload(args));
       },
       // Log a warning. The default implementation wraps
       // `console.warn` and passes on all arguments,
@@ -554,8 +584,8 @@ module.exports = {
       // The intention is that `apos.util.warn` should be
       // called for situations less dire than
       // `apos.util.error`.
-      warn(msg) {
-        self.logger.warn.apply(self.logger, arguments);
+      warn(...args) {
+        self.logger.warn(...self.convertLegacyLogPayload(args));
       },
 
       // Identical to `apos.util.warn`, except that the warning is
@@ -605,8 +635,8 @@ module.exports = {
       // Overrides should be written with support for
       // substitution strings in mind. See the
       // `console.error` documentation.
-      error(msg) {
-        self.logger.error.apply(self.logger, arguments);
+      error(...args) {
+        self.logger.error(...self.convertLegacyLogPayload(args));
       },
       // Performance profiling method. At the start of the operation you want
       // to profile, call with req (may be null or omitted entirely) and a
@@ -811,6 +841,46 @@ module.exports = {
       },
       omit(source, keys) {
         return _.omit(source, keys);
+      },
+
+      // Internal method. Attempt to convert the log payload to an object
+      // for legacy calls and when `@apostrophecms/log` has been configured
+      // with `messageAs: 'someKey'`.
+      // This change is backwards compatible with the previous behavior because
+      // `messageAs` is a newly introduced option. Custom loggers should adapt
+      // to this change when using `messageAs`.
+      // `args` is the argument array passed to the any log method.
+      // The result (when required) is an array with a single object.
+      // First string argument (if available) is used as `message`.
+      // First object argument is used as result object.
+      // All other arguments are passed as `args` property of the result object.
+      convertLegacyLogPayload(args) {
+        const messageAs = self.apos.structuredLog.options.messageAs;
+        if (!messageAs || args.length === 0) {
+          return args;
+        }
+        // Already formatted by the structured log module. Nothing we can do if not.
+        if (args.length === 1 && _.isPlainObject(args[0])) {
+          return args;
+        }
+
+        // Should also handle apos.util.warnDev() calls.
+        const messageIndex = args
+          .findIndex(arg => typeof arg === 'string' && arg.trim() && arg !== '\n⚠️ ');
+        const firstObjectIndex = args.findIndex(arg => _.isPlainObject(arg));
+        const message = messageIndex !== -1 ? args[messageIndex] : null;
+        const firstObject = firstObjectIndex !== -1 ? { ...args[firstObjectIndex] } : {};
+
+        if (message) {
+          firstObject[messageAs] = message;
+        }
+        const rest = args
+          .filter((arg, index) => ![ messageIndex, firstObjectIndex ].includes(index));
+        if (rest.length) {
+          firstObject.args = rest;
+        }
+
+        return [ firstObject ];
       }
     };
   },